Add files via upload
In this we use Feature as a rating which are given by the Employee, salary of the Employee, was he/she promoted or not during last 5 years, how much they are spending time in the company, how many projects they are done, on the basis of this basic feature we first show the correlation between them and target column then made two Models one is XGBClassifier and another one is RandomForestClassifier. After testing both the model prediction we see that XGBClassifier prediction is 99.16% which is extremely good then we then fit the model again with XGBClassifier and predict the Target value on the basis of this.